Support specific branches, tags, and commits
Using the syntax owner/repo@ref, you can now specify a branch, tag, or
commit. We continue to default to the repository's default branch.

When an explicit ref is requested, it is included in the output.

"repo spec" parsing has also moved to the cmd layer, which is a better
fit for multiple reasons.